examples/quickstart-tensorflow-phe/pyproject.toml
# pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name        = "pheexample"
version     = "1.0.0"
description = "Flower + TensorFlow con HE Paillier (python‑phe) vs TenSEAL"
license     = "Apache-2.0"
dependencies = [
  "flwr[simulation]==1.19.0",
  "flwr-datasets[vision]==0.4.0",
  "tensorflow-cpu==2.15.0 ; platform_machine=='x86_64'",
  "tensorflow-macos>=2.9.1,!=2.11.1 ; platform_machine=='arm64'",
  "phe>=1.4.0",             # Paillier
  "matplotlib",
  "pandas",
]

[tool.hatch.build.targets.wheel]
packages = ["pheexample"]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "pheexample.server_app:app"
clientapp = "pheexample.client_app:app"


[tool.flwr.app."quickstart-tensorflow-phe/pheexample/server_app.py:app"]
client-app = "quickstart-tensorflow-phe/pheexample/client_app.py:app"
num-supernodes = 10
run-config = { num-server-rounds = 20, local-epochs = 1, batch-size = 32, learning-rate = 0.005, fraction-fit = 0.5, fraction-evaluate = 1.0, verbose = false, seed = 44 }

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 10
